remove casting of severeCasesByRequestedTime

+++ src/estimator.py
@@ -31,23 +31,18 @@ def severe_cases_by_time(infections):
     This is the estimated number of severe positive
     cases that will require hospitalization to recover.
     """
-
-    return int(infections*15/100)
+    return infections*15/100
 
 
 def hospital_beds_by_time(beds_count, infections):
     available_beds = int(beds_count * 35/100)
     return available_beds - infections
 
-
-
 def icu_request_by_time(infections):
     return int(infections * 5/100)
 
-
 def ventilators_request_by_time(infections):
     return int(infections * 2/100)
-
 
 def dollars_in_flight(infections, avg_daily_income, time_to_elapse, period_type='days'):
